11-13-2

+++ src/main/java/com/suma/controller/BootAdvController.java
@@ -5,11 +5,13 @@ import com.google.common.collect.Lists;
 import com.suma.constants.AdvContants;
 import com.suma.constants.ExceptionConstants;
 import com.suma.exception.AdvInfoException;
+import com.suma.exception.AdvMaterialException;
 import com.suma.pojo.*;
 import com.suma.service.*;
 import com.suma.service.impl.AdvRegionService;
 import com.suma.utils.Result;
 import com.suma.utils.StringUtil;
+import com.suma.utils.UserAndTimeUtils;
 import com.suma.vo.BootAdvVO;
 import com.suma.vo.InfoMaterialVO;
 import org.apache.commons.lang3.StringUtils;
@@ -53,26 +55,45 @@ public class BootAdvController extends BaseController {
         if (Objects.isNull(bootAdvVO)) {
             throw new AdvInfoException(ExceptionConstants.INFO_EXCEPTION_MISSING_REQUIRED_PARAMS);
         }
+        Boolean flag = bootAdvVO.getAdvTypeId().equals(AdvContants.START_LOGO_ADV_LOCATION_ID)
+                || bootAdvVO.getAdvTypeId().equals(AdvContants.START_MACHINE_ADV_SUBTYPE_ID);
+        if (!flag) {
+            throw new AdvInfoException("请添加开机广告");
+        }
         try {
             //读取广告信息
             AdvInfo advInfo = new AdvInfo();
             BeanUtils.copyProperties(bootAdvVO, advInfo);
-            advInfo.setAdvLocationId(AdvContants.START_ADV_LOCATION_ID);
+            //判断广告信息是否存在
+            advInfoService.judgeAdvInfo(advInfo);
+            advInfo = UserAndTimeUtils.setCreateUserAndTime(advInfo);
+            //设置开机广告位
+            advInfoService.setBootLocation(advInfo);
             advInfoService.save(advInfo);
             //获取广告信息ID
             Long advInfoId = advInfo.getId();
-            //设置开机广告位
-            advInfo.setAdvLocationId(AdvContants.START_LOGO_ADV_LOCATION_ID);
             //保存广告版本号
             //获取并保存区域信息
-            List<Integer> regionIds = bootAdvVO.getRegionIds();
+            List<Integer> regionIds = bootAdvVO.getRegionId();
             infoRegionService.saveInfoRegion(regionIds, advInfoId);
             //获取广告资源信息
-            infoMaterialService.saveInfoMaterial(bootAdvVO.getInfoMaterialVOS(), advInfoId);
+            List<InfoMaterialVO> infoMaterialVOS = bootAdvVO.getInfoMaterialVOS();
+            if (bootAdvVO.getMaterialType().equals(AdvContants.VEDIO_MATERIAL) & !CollectionUtils.isEmpty(infoMaterialVOS)) {
+                if (infoMaterialVOS.size() > 1) {
+                    throw new AdvMaterialException("开机视频广告仅支持单个视频");
+                }
+            }
+            if (bootAdvVO.getMaterialType().equals(AdvContants.IMAGE_MATERIAL) & !CollectionUtils.isEmpty(infoMaterialVOS)) {
+                if (infoMaterialVOS.size() > 5) {
+                    throw new AdvMaterialException("开机图片广告最多可配置5张图片");
+                }
+            }
+            infoMaterialService.saveInfoMaterial(infoMaterialVOS, advInfoId);
             //保存软件版本号
             infoVersionService.saveVersion(advInfoId,bootAdvVO.getVersion());
         } catch (Exception e) {
-            return Result.error();
+            e.printStackTrace();
+            throw e;
         }
         return Result.success();
     }
@@ -89,23 +110,30 @@ public class BootAdvController extends BaseController {
         if (Objects.isNull(bootAdvVO)) {
             throw new AdvInfoException(ExceptionConstants.INFO_EXCEPTION_MISSING_REQUIRED_PARAMS);
         }
+        Boolean flag = bootAdvVO.getAdvTypeId().equals(AdvContants.START_LOGO_ADV_LOCATION_ID)
+                || bootAdvVO.getAdvTypeId().equals(AdvContants.START_MACHINE_ADV_SUBTYPE_ID);
+        if (!flag) {
+            throw new AdvInfoException("不是开机广告类型");
+        }
         try {
             //获取广告信息
             AdvInfo advInfo = new AdvInfo();
             BeanUtils.copyProperties(bootAdvVO, advInfo);
+            advInfo = UserAndTimeUtils.setEditUserAndTime(advInfo);
             advInfoService.updateByPrimaryKeySelective(advInfo);
             Long advInfoId = advInfo.getId();
             //删除原来的区域信息
             infoRegionService.deleteByAdvInfoId(advInfoId);
             //获取区域信息,并保存
-            List<Integer> regionIds = bootAdvVO.getRegionIds();
+            List<Integer> regionIds = bootAdvVO.getRegionId();
             infoRegionService.saveInfoRegion(regionIds, advInfoId);
             //删除原来资源信息
             infoMaterialService.deleteByAdvInfoId(advInfoId);
             //获取资源对应信息
             infoMaterialService.saveInfoMaterial(bootAdvVO.getInfoMaterialVOS(), advInfoId);
         } catch (Exception e) {
-            return Result.error();
+            e.printStackTrace();
+            throw e;
         }
         return Result.success();
     }
@@ -125,39 +153,15 @@ public class BootAdvController extends BaseController {
         try {
             //将数组转换成list集合
             List<Long> advInfoIds = StringUtil.convertstr(str);
-            //删除广告信息
-            advInfoService.deleteByAdvInfoIds(advInfoIds);
-            //删除广告对应资源信息
-            infoMaterialService.deleteByAdvInfoIds(advInfoIds);
-            //删除广告对应区域信息
-            infoRegionService.deleteByAdvInfoIds(advInfoIds);
-            //删除广告位
-            deleteAdvLocation(advInfoIds);
-            //删除广告对应的所有版本信息
-            infoVersionService.deleteByAdvInfoIds(advInfoIds,null);
+            if (!CollectionUtils.isEmpty(advInfoIds)) {
+                advInfoService.deleteAdvRelationInfo(advInfoIds);
+                infoVersionService.deleteByAdvInfoIds(advInfoIds,null);
+            }
         } catch (Exception e) {
-            return Result.error();
+            e.printStackTrace();
+            throw e;
         }
         return Result.success();
-    }
-    /**
-     * 删除广告位
-     * @param advInfoIds
-     */
-    public void deleteAdvLocation(List<Long> advInfoIds) {
-        List<Long> list = Lists.newArrayList();
-        if (!CollectionUtils.isEmpty(advInfoIds)) {
-            advInfoIds.forEach(id ->{
-                if (advInfoService.findById(id) != null) {
-                    list.add(advInfoService.findById(id).getAdvLocationId());
-                }
-            });
-        }
-        if (!CollectionUtils.isEmpty(list)) {
-            AdvLocationExample example = new AdvLocationExample();
-            example.createCriteria().andIdIn(list);
-            advLocationService.deleteByExample(example);
-        }
     }
 
     /**
@@ -193,7 +197,7 @@ public class BootAdvController extends BaseController {
         //所有区域
         bootAdvVO.setAdvRegions(advRegionService.selectAdvRegionAll());
         //有效区域ID
-        bootAdvVO.setRegionIds(infoRegionService.getRegionIds(advInfo.getId()));
+        bootAdvVO.setRegionId(infoRegionService.getRegionIds(advInfo.getId()));
         return Result.success(bootAdvVO);
     }
 }
